[sensors] move Temperatures from host to sensors.
