+++ host/host_freebsd.go
@@ -141,10 +141,6 @@ func getUsersFromUtmp(utmpfile string) ([]UserStat, error) {
 	return ret, nil
 }
 
-func SensorsTemperaturesWithContext(ctx context.Context) ([]TemperatureStat, error) {
-	return []TemperatureStat{}, common.ErrNotImplementedError
-}
-
 func KernelVersionWithContext(ctx context.Context) (string, error) {
 	_, _, version, err := PlatformInformationWithContext(ctx)
 	return version, err
