+++ host/host_solaris.go
@@ -1,13 +1,13 @@
 // SPDX-License-Identifier: BSD-3-Clause
+//go:build solaris
+
 package host
 
 import (
 	"bufio"
 	"bytes"
 	"context"
-	"encoding/csv"
 	"fmt"
-	"io"
 	"os"
 	"regexp"
 	"strconv"
@@ -93,43 +93,6 @@ func UptimeWithContext(ctx context.Context) (uint64, error) {
 
 func UsersWithContext(ctx context.Context) ([]UserStat, error) {
 	return []UserStat{}, common.ErrNotImplementedError
-}
-
-func SensorsTemperaturesWithContext(ctx context.Context) ([]TemperatureStat, error) {
-	var ret []TemperatureStat
-
-	out, err := invoke.CommandWithContext(ctx, "ipmitool", "-c", "sdr", "list")
-	if err != nil {
-		return ret, err
-	}
-
-	r := csv.NewReader(strings.NewReader(string(out)))
-	// Output may contain errors, e.g. "bmc_send_cmd: Permission denied", don't expect a consistent number of records
-	r.FieldsPerRecord = -1
-	for {
-		record, err := r.Read()
-		if err == io.EOF {
-			break
-		}
-		if err != nil {
-			return ret, err
-		}
-		// CPU1 Temp,40,degrees C,ok
-		if len(record) < 3 || record[1] == "" || record[2] != "degrees C" {
-			continue
-		}
-		v, err := strconv.ParseFloat(record[1], 64)
-		if err != nil {
-			return ret, err
-		}
-		ts := TemperatureStat{
-			SensorKey:   strings.TrimSuffix(record[0], " Temp"),
-			Temperature: v,
-		}
-		ret = append(ret, ts)
-	}
-
-	return ret, nil
 }
 
 func VirtualizationWithContext(ctx context.Context) (string, string, error) {
