+++ host/host_windows.go
@@ -6,7 +6,6 @@ package host
 import (
 	"context"
 	"fmt"
-	"math"
 	"strconv"
 	"strings"
 	"sync/atomic"
@@ -16,7 +15,6 @@ import (
 
 	"github.com/shirou/gopsutil/v4/internal/common"
 	"github.com/shirou/gopsutil/v4/process"
-	"github.com/yusufpapurcu/wmi"
 	"golang.org/x/sys/windows"
 )
 
@@ -55,13 +53,6 @@ type systemInfo struct {
 	dwAllocationGranularity     uint32
 	wProcessorLevel             uint16
 	wProcessorRevision          uint16
-}
-
-type msAcpi_ThermalZoneTemperature struct {
-	Active             bool
-	CriticalTripPoint  uint32
-	CurrentTemperature uint32
-	InstanceName       string
 }
 
 func HostIDWithContext(ctx context.Context) (string, error) {
@@ -230,33 +221,6 @@ func UsersWithContext(ctx context.Context) ([]UserStat, error) {
 	var ret []UserStat
 
 	return ret, common.ErrNotImplementedError
-}
-
-func SensorsTemperaturesWithContext(ctx context.Context) ([]TemperatureStat, error) {
-	var ret []TemperatureStat
-	var dst []msAcpi_ThermalZoneTemperature
-	q := wmi.CreateQuery(&dst, "")
-	if err := common.WMIQueryWithContext(ctx, q, &dst, nil, "root/wmi"); err != nil {
-		return ret, err
-	}
-
-	for _, v := range dst {
-		ts := TemperatureStat{
-			SensorKey:   v.InstanceName,
-			Temperature: kelvinToCelsius(v.CurrentTemperature, 2),
-		}
-		ret = append(ret, ts)
-	}
-
-	return ret, nil
-}
-
-func kelvinToCelsius(temp uint32, n int) float64 {
-	// wmi return temperature Kelvin * 10, so need to divide the result by 10,
-	// and then minus 273.15 to get °Celsius.
-	t := float64(temp/10) - 273.15
-	n10 := math.Pow10(n)
-	return math.Trunc((t+0.5/n10)*n10) / n10
 }
 
 func VirtualizationWithContext(ctx context.Context) (string, string, error) {
