+++ sensors/sensors_darwin_cgo.go
@@ -1,14 +1,14 @@
 // SPDX-License-Identifier: BSD-3-Clause
 //go:build darwin && cgo
 
-package host
+package sensors
 
 // #cgo LDFLAGS: -framework IOKit
 // #include "smc_darwin.h"
 import "C"
 import "context"
 
-func SensorsTemperaturesWithContext(ctx context.Context) ([]TemperatureStat, error) {
+func TemperaturesWithContext(ctx context.Context) ([]TemperatureStat, error) {
 	temperatureKeys := []string{
 		C.AMBIENT_AIR_0,
 		C.AMBIENT_AIR_1,
