+++ sensors/sensors_darwin_nocgo.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: BSD-3-Clause
 //go:build darwin && !cgo
 
-package host
+package sensors
 
 import (
 	"context"
@@ -9,6 +9,6 @@ import (
 	"github.com/shirou/gopsutil/v4/internal/common"
 )
 
-func SensorsTemperaturesWithContext(ctx context.Context) ([]TemperatureStat, error) {
+func TemperaturesWithContext(ctx context.Context) ([]TemperatureStat, error) {
 	return []TemperatureStat{}, common.ErrNotImplementedError
 }
