+++ sensors/smc_darwin.c
