+++ sensors/smc_darwin.h
